+++ RevitAddin.CommandBindingTester/Views/BindingView.xaml.cs
@@ -11,36 +11,8 @@ namespace RevitAddin.CommandBindingTester.Views
         {
             InitializeComponent();
             InitializeWindow();
-            UpdateDataGridWithStyle(dataGrid);
+            dataGrid.UpdateDataGridStyle();
             Title = "RevitCommand Binding Viewer";
-        }
-
-        private void UpdateDataGridWithStyle(DataGrid dataGrid)
-        {
-            var brush = new SolidColorBrush();
-            brush.Color = Color.FromRgb(211, 211, 211); //#FFD3D3D3
-            brush.Opacity = 0.3;
-            dataGrid.HorizontalGridLinesBrush = brush;
-            dataGrid.VerticalGridLinesBrush = brush;
-
-            dataGrid.SelectionMode = DataGridSelectionMode.Extended;
-            dataGrid.SelectionUnit = DataGridSelectionUnit.Cell;
-
-            var alternativeBrush = new SolidColorBrush();
-            alternativeBrush.Color = Color.FromRgb(211, 211, 211); //#FFD3D3D3
-            alternativeBrush.Opacity = 0.1;
-            dataGrid.AlternatingRowBackground = alternativeBrush;
-
-            dataGrid.CanUserAddRows = false;
-
-            dataGrid.CanUserReorderColumns = true;
-            dataGrid.CanUserResizeColumns = true;
-            dataGrid.CanUserSortColumns = true;
-            dataGrid.CanUserResizeRows = false;
-
-            dataGrid.AutoGenerateColumns = true;
-
-            dataGrid.HeadersVisibility = DataGridHeadersVisibility.Column;
         }
 
         #region InitializeWindow