+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.0] / 2025-10-04
 ### Features
 - Binding Viewer with all available `PostableCommand` with binding information.
+### Updates
+- Add `DataGridStyleExtension.UpdateDataGridStyle` with auto generated columns size.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0